requirements.txt
fastapi
uvicorn
psycopg2-binary
redis
rq
requests
python-dotenv
pydantic
google-api-python-client
google-auth
google-auth-httplib2
